# Add StringMap and use it instead of LablesMap