--- a/internal/data/common.go
+++ b/internal/data/common.go
@@ -18,6 +18,8 @@ package data
 // such as timestamps, attributes, etc.
 
 import (
+	"sort"
+
 	otlpcommon "github.com/open-telemetry/opentelemetry-proto/gen/go/common/v1"
 )
 
@@ -138,6 +140,135 @@ type Attributes struct {
 
 func NewAttributes(m AttributesMap, droppedCount uint32) Attributes {
 	return Attributes{m, droppedCount}
+}
+
+// StringKeyValue stores a key and value pair.
+type StringKeyValue struct {
+	orig *otlpcommon.StringKeyValue
+}
+
+// NewStringKeyValue creates a new StringKeyValue with the given key and value.
+func NewStringKeyValue(k string, v string) StringKeyValue {
+	return StringKeyValue{&otlpcommon.StringKeyValue{Key: k, Value: v}}
+}
+
+// Key returns the key associated with this StringKeyValue.
+func (akv StringKeyValue) Key() string {
+	return akv.orig.Key
+}
+
+// Value returns the value associated with this StringKeyValue.
+func (akv StringKeyValue) Value() string {
+	return akv.orig.Value
+}
+
+// SetValue replaces the value associated with this StringKeyValue.
+func (akv StringKeyValue) SetValue(v string) {
+	akv.orig.Value = v
+}
+
+// StringMap stores a map of attribute keys to values.
+type StringMap struct {
+	orig *[]*otlpcommon.StringKeyValue
+}
+
+func newStringMap(orig *[]*otlpcommon.StringKeyValue) StringMap {
+	return StringMap{orig}
+}
+
+// NewStringMap creates a new StringMap from the given map[string]string.
+func NewStringMap(attrMap map[string]string) StringMap {
+	if len(attrMap) == 0 {
+		var orig []*otlpcommon.StringKeyValue
+		return StringMap{&orig}
+	}
+	origs := make([]otlpcommon.StringKeyValue, len(attrMap))
+	wrappers := make([]*otlpcommon.StringKeyValue, len(attrMap))
+
+	ix := 0
+	for k, v := range attrMap {
+		wrappers[ix] = &origs[ix]
+		wrappers[ix].Key = k
+		wrappers[ix].Value = v
+		ix++
+	}
+
+	return StringMap{&wrappers}
+}
+
+// Get returns the StringKeyValue associated with the key and true,
+// otherwise an invalid instance of the StringKeyValue and false.
+func (sm StringMap) Get(k string) (StringKeyValue, bool) {
+	for _, a := range *sm.orig {
+		if a.Key == k {
+			return StringKeyValue{a}, true
+		}
+	}
+	return StringKeyValue{nil}, false
+}
+
+// Delete deletes the entry associated with the key and returns true if the key
+// was present in the map, otherwise returns false.
+func (sm StringMap) Delete(k string) bool {
+	for i, a := range *sm.orig {
+		if a.Key == k {
+			(*sm.orig)[i] = (*sm.orig)[len(*sm.orig)-1]
+			*sm.orig = (*sm.orig)[:len(*sm.orig)-1]
+			return true
+		}
+	}
+	return false
+}
+
+// Insert adds the StringKeyValue to the map when the key does not exist.
+// No action is applied to the map where the key already exists.
+func (sm StringMap) Insert(k, v string) {
+	if _, existing := sm.Get(k); !existing {
+		*sm.orig = append(*sm.orig, NewStringKeyValue(k, v).orig)
+	}
+}
+
+// Update updates an existing StringKeyValue with a value.
+// No action is applied to the map where the key does not exist.
+func (sm StringMap) Update(k, v string) {
+	if av, existing := sm.Get(k); existing {
+		av.SetValue(v)
+	}
+}
+
+// Upsert performs the Insert or Update action. The StringKeyValue is
+// insert to the map that did not originally have the key. The key/value is
+// updated to the map where the key already existed.
+func (sm StringMap) Upsert(k, v string) {
+	if av, existing := sm.Get(k); existing {
+		av.SetValue(v)
+	} else {
+		*sm.orig = append(*sm.orig, NewStringKeyValue(k, v).orig)
+	}
+}
+
+// Len returns the number of StringKeyValue in the map.
+func (sm StringMap) Len() int {
+	return len(*sm.orig)
+}
+
+// GetStringKeyValue returns the StringKeyValue associated with the given index.
+//
+// This function is used mostly for iterating over all the values in the map:
+// for i := 0; i < am.Len(); i++ {
+//     akv := am.GetStringKeyValue(i)
+//     ... // Do something with the attribute
+// }
+func (sm StringMap) GetStringKeyValue(ix int) StringKeyValue {
+	return StringKeyValue{(*sm.orig)[ix]}
+}
+
+// Sort sorts the entries in the StringMap so two instances can be compared.
+// Returns the same instance to allow nicer code like:
+// assert.EqualValues(t, expected.Sort(), actual.Sort())
+func (sm StringMap) Sort() StringMap {
+	sort.SliceStable(*sm.orig, func(i, j int) bool { return (*sm.orig)[i].Key < (*sm.orig)[j].Key })
+	return sm
 }
 
 // InstrumentationLibrary is a message representing the instrumentation library information.